+++ README.md
@@ -1,1 +1,1 @@
-# coda
+# kodit
