--- src/coda/__init__.py
@@ -1,1 +1,0 @@
-"""Coda - Code indexing for better AI code generation."""
