+++ src/kodit/__init__.py
@@ -1,0 +1,1 @@
+"""kodit - Code indexing for better AI code generation."""
